fix - Enabled SSTAFAnalyzerIntegrationTest and fixed test failures
Addresses SSTAF Issue 6 - Exception in SSTAF Analyzer

SSTAFAnalyzerIntegrationTest was not firing during the build due to a problem in the build.gradle file. Because the test didn't run, problems with the Analyzer's I/O were not discovered. The tests now run and the Analyzer appears robust. The analyzer can now ignore a finite number of incoming nulls and can accumulate a single message across nulls and newlines.

This commit builds on Mac. Windows will be checked next.