enhancement - 4-create-example-features-and-application
Added an example of implementing tick() in an agent that is backed by a python model.

Added a static Address that routes messages to the client. This enables Features to send messages to the client without having to get the client address into the feature. Message routing in the EntityController was modified to support the address.